small work

+++ src/main/java/cn/warriorView/Util/MathUtil.java
@@ -5,8 +5,8 @@ import java.math.RoundingMode;
 
 public class MathUtil {
 
-    public static double round(double num, int decimalPlaces) {
-        return new BigDecimal(num).setScale(decimalPlaces, RoundingMode.HALF_UP).doubleValue();
+    public static float round(Float num, int decimalPlaces) {
+        return new BigDecimal(num).setScale(decimalPlaces, RoundingMode.HALF_UP).floatValue();
     }
 
 }
